Fixed 'next' and 'previous' buttons
Next and previous both use a start_slice variable to subset the dataframe, but the radial button selection code did not. This resulted in radial button selection not matching the displayed items when using 'next' or 'previous'. Now the start_slice variable is added to the dataframe subset when the radial button is pressed.